requirement.txt
absl-py==0.12.0
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.3.0
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.5
channels==3.0.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
coreschema==0.0.4
cryptography==3.4.7
cupshelpers==1.0
cycler==0.10.0
daphne==3.0.2
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.4
django-crispy-forms==1.11.2
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
inflection==0.5.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==21.1.0
simplejson==3.16.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zope.interface==5.4.0
